+++ R/group-data-SpatVector.R
@@ -34,7 +34,7 @@
 #'
 #' @return
 #'
-#' See the description of the function. The results are usually tibbles,
+#' See the description of the method. The results are usually tibbles,
 #' lists or vectors. These functions does not return `SpatVector` objects.
 #'
 #' @rdname group_data.SpatVector