+++ R/mutate-Spat.R
@@ -5,12 +5,6 @@
 #' `mutate()` adds new layers/attributes and preserves existing ones on a
 #' `Spat*` object.
 #'
-#' `r lifecycle::badge("superseded")`
-#'
-#' `transmute()` creates a new object containing only the specified
-#' computations. It's superseded because you can perform the same job
-#' with `mutate(.keep = "none")`.
-#'
 #' @inherit select.Spat return
 #'
 #' @param .data A `SpatRaster` created with [terra::rast()] or a `SpatVector`
@@ -22,13 +16,11 @@
 #' @rdname mutate.Spat
 #' @name mutate.Spat
 #'
-#' @aliases transmute.Spat
-#'
 #' @importFrom dplyr mutate
 #'
 #' @seealso
 #'
-#' [dplyr::mutate()], [dplyr::transmute()] methods.
+#' [dplyr::mutate()] methods.
 #'
 #' \CRANpkg{terra} provides several ways to modify `Spat*` objects:
 #'
@@ -49,8 +41,7 @@
 #'
 #' @section Methods:
 #'
-#' Implementation of the **generic** [dplyr::mutate()], [dplyr::transmute()]
-#' functions.
+#' Implementation of the **generic** [dplyr::mutate()] method.
 #'
 #' ## `SpatRaster`
 #'
@@ -58,14 +49,10 @@
 #' `SpatRaster` with the same extent, resolution and CRS than `.data`. Only the
 #' values (and possibly the number) of layers is modified.
 #'
-#' `transmute()` would keep only the layers created with `...`.
-#'
 #' ## `SpatVector`
 #'
 #' The result is a `SpatVector` with the modified (and possibly renamed)
 #' attributes on the function call.
-#'
-#' `transmute()` would keep only the attributes created with `...`.
 #'
 #' @examples
 #'
@@ -165,9 +152,8 @@ mutate.SpatVector <- function(
   )
 
   # If NULL...
-  if(ncol(mutated) == 0){
-    vend <- .data[,0]
-
+  if (ncol(mutated) == 0) {
+    vend <- .data[, 0]
   } else {
     # Bind
     vend <- cbind(.data[, 0], mutated)
@@ -178,91 +164,6 @@ mutate.SpatVector <- function(
 
   vend
 }
-#' @export
-#' @rdname mutate.Spat
-#' @importFrom dplyr transmute
-transmute.SpatRaster <- function(.data, ...) {
-  df <- as_tbl_internal(.data)
-
-  xy <- dplyr::select(df, 1:2)
-
-  values <- dplyr::select(df, -c(1, 2))
-
-  values_transm <- dplyr::transmute(values, ...)
-
-  # dtplyr
-  xy <- data.table::as.data.table(xy)
-  values_transm <- data.table::as.data.table(values_transm)
-
-  final_df <- dplyr::bind_cols(xy, values_transm)
-
-  # To data.table and rearrange attrs
-  final_df <- data.table::as.data.table(final_df)
-
-  # Spatial attrs
-  init_att <- attributes(df)
-  final_att <- attributes(final_df)
-
-  spat_attrs <- init_att[setdiff(names(init_att), names(final_att))]
-
-  attributes(final_df) <- c(final_att, spat_attrs)
-
-  # Rearrange number of layers
-  dims <- attributes(df)$dims
-  dims[3] <- ncol(values_transm)
-  attr(final_df, "dims") <- dims
-
-  final_rast <- as_spat_internal(final_df)
-
-  # Check coltab
-  if (
-    any(terra::has.colors(.data)) && any(names(final_rast) %in% names(.data))
-  ) {
-    ctab_list_init <- terra::coltab(.data)
-    names(ctab_list_init) <- names(.data)
-    namesend <- names(final_rast)
-
-    ctab_list <- ctab_list_init[namesend %in% names(.data)]
-
-    # Assign coltab by layer
-    l2 <- lapply(seq_len(terra::nlyr(final_rast)), function(x) {
-      rr <- terra::subset(final_rast, x)
-      if (names(rr) %in% names(ctab_list)) {
-        ctab <- ctab_list[match(names(rr), names(ctab_list))]
-      } else {
-        ctab <- NULL
-      }
-
-      terra::coltab(rr) <- ctab
-      rr
-    })
-    final_rast <- do.call("c", l2)
-  }
-
-  final_rast
-}
-#' @export
-#' @rdname mutate.Spat
-transmute.SpatVector <- function(.data, ...) {
-  # Use own method
-  tbl <- as_tibble(.data)
-  transm <- dplyr::transmute(tbl, ...)
-
-  if (ncol(transm) > 0) {
-    # Bind
-    vend <- cbind(.data[, 0], transm)
-  } else {
-    vend <- .data[, 0]
-  }
-
-  # Prepare groups
-  vend <- group_prepare_spat(vend, transm)
-
-  vend
-}
 
 #' @export
 dplyr::mutate
-
-#' @export
-dplyr::transmute